Introduce HttpClient.Metrics and HttpClient.Builder.State
